Move SAM2 torch cache tests into inference_models
